requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
daphne==4.1.2
Django==4.2.11
django-channels==0.7.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.7
importlib-resources==6.4.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.3.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.7
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
txaio==23.1.1
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.18.1
zope.interface==6.3